Move serializing to one file for common
#833
