+++ trview.common/Colour.cpp
@@ -87,27 +87,4 @@ namespace trview
     {
         return left.a == right.a && left.r == right.r && left.g == right.g && left.b == right.b;
     }
-
-    void from_json(const nlohmann::json& json, Colour& colour)
-    {
-        if (json.is_string())
-        {
-            const auto str = json.get<std::string>();
-            if (str == "transparent")
-            {
-                colour = Colour::Transparent;
-            }
-            else
-            {
-                colour = Colour::White;
-            }
-        }
-        else
-        {
-            colour = Colour(read_attribute<float>(json, "a", 1.0f),
-                read_attribute<float>(json, "r", 0.0f),
-                read_attribute<float>(json, "g", 0.0f),
-                read_attribute<float>(json, "b", 0.0f));
-        }
-    }
 }
