+++ trview.common/Colour.h
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <SimpleMath.h>
-#include <external/nlohmann/json.hpp>
 
 namespace trview
 {
@@ -51,6 +50,4 @@ namespace trview
     Colour operator+(const Colour& left, const Colour& right);
 
     bool operator==(const Colour& left, const Colour& right);
-
-    void from_json(const nlohmann::json& json, Colour& colour);
 }
