+++ trview.common/Point.cpp
@@ -40,9 +40,4 @@ namespace trview
         return x >= first.x && y >= first.y 
             && x <= second.x && y <= second.y; 
     }
-
-    void from_json(const nlohmann::json& json, Point& point)
-    {
-        point = Point(read_attribute<float>(json, "x"), read_attribute<float>(json, "y"));
-    }
 }
