+++ trview.common/Point.h
@@ -1,7 +1,5 @@
 #pragma once
 
-#include <external/nlohmann/json.hpp>
-
 namespace trview
 {
     struct Point
@@ -19,6 +17,4 @@ namespace trview
 
         float x, y;
     };
-
-    void from_json(const nlohmann::json& json, Point& point);
 }
