+++ trview.common/Size.cpp
@@ -44,9 +44,4 @@ namespace trview
     {
         return !operator==(size);
     }
-
-    void from_json(const nlohmann::json& json, Size& size)
-    {
-        size = Size(read_attribute<float>(json, "width"), read_attribute<float>(json, "height"));
-    }
 }
