+++ trview.common/Size.h
@@ -4,7 +4,6 @@
 /// A floating point based two dimensional size.
 
 #pragma once
-#include <external/nlohmann/json.hpp>
 
 namespace trview
 {
@@ -49,6 +48,4 @@ namespace trview
 
         float width, height;
     };
-
-    void from_json(const nlohmann::json& json, Size& size);
 }
